Basic refactoring to make it easier to walk through the requirements
