+++ Sources/SKTiledKit/SpriteKit/Level Processing/SKScene+TK.swift
@@ -1,0 +1,31 @@
+//    Copyright 2020 Swift Studies
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
+import SpriteKit
+import TiledKit
+
+extension SKScene : SpecializedLevel {
+    public typealias Container = SKNode
+    
+    public var primaryContainer: Container {
+        return self
+    }
+
+    public func apply(tiledLevel level: Level) {
+        scene?.size =  CGSize(width: CGFloat(level.width*level.tileWidth), height: CGFloat(level.height*level.tileHeight))
+        scene?.userData = NSMutableDictionary()
+
+        apply(propertiesFrom: level)
+    }
+}
