+++ Sources/SKTiledKit/SpriteKit/Level+SKTK.swift
@@ -1,0 +1,28 @@
+//    Copyright 2020 Swift Studies
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
+import TiledKit
+import SpriteKit
+
+extension Level {
+    var heightInPixels : CGFloat {
+        return (height * tileHeight).cgFloatValue
+    }
+    
+    func transform(y:Double, objectHeight height:Double = 0)->CGFloat{
+        let levelHeight = Double(self.height) * Double(self.tileHeight)
+        
+        return CGFloat((levelHeight - y) - (height / 2))
+    }
+}
